Bug fix for long parser

+++ common/src/main/java/cc/blynk/common/utils/ParseUtil.java
@@ -17,7 +17,7 @@ public final class ParseUtil {
 
     public static long parseLong(String longProperty) {
         try {
-            return Integer.parseInt(longProperty);
+            return Long.parseLong(longProperty);
         } catch (NumberFormatException nfe) {
             throw new RuntimeException(longProperty + " not a number. " + nfe.getMessage());
         }
